+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.21.0](https://github.com/elephant-xyz/elephant-cli/compare/v1.20.1...v1.21.0) (2025-08-05)
+
+
+### Features
+
+* **cli:** add check-transaction-status command ([#72](https://github.com/elephant-xyz/elephant-cli/issues/72)) ([c1c82ae](https://github.com/elephant-xyz/elephant-cli/commit/c1c82aec63b357ee22d1f731641dc9596bc06bb7))
+
+## [1.20.1](https://github.com/elephant-xyz/elephant-cli/compare/v1.20.0...v1.20.1) (2025-08-05)
+
+
+### Bug Fixes
+
+* **json-validator:** allow zero as valid currency value ([#71](https://github.com/elephant-xyz/elephant-cli/issues/71)) ([fd6d01d](https://github.com/elephant-xyz/elephant-cli/commit/fd6d01d30ce716c6f910c6b333a996b70584589c))
+
 # [1.20.0](https://github.com/elephant-xyz/elephant-cli/compare/v1.19.0...v1.20.0) (2025-08-05)